layout improvements
