+++ components/tactility/src/apps/system/wifi/wifi.h
@@ -1,8 +1,11 @@
 #pragma once
-#include "services/wifi/wifi.h"
 #include "mutex.h"
-#include "wifi_state.h"
+#include "services/wifi/wifi.h"
 #include "wifi_view.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
 
 typedef struct {
     FuriPubSubSubscription* wifi_subscription;
@@ -14,3 +17,7 @@ typedef struct {
 void wifi_lock(Wifi* wifi);
 
 void wifi_unlock(Wifi* wifi);
+
+#ifdef __cplusplus
+}
+#endif
