+++ components/tactility/src/apps/system/wifi/wifi_connect_view.c
@@ -1,0 +1,16 @@
+#include "wifi_connect_view.h"
+
+#include "lvgl.h"
+#include "wifi_state.h"
+
+#define TAG "wifi_connect_view"
+
+void wifi_connect_view_create(WifiConnectView* view, lv_obj_t* parent) {
+}
+
+void wifi_connect_view_update(WifiConnectView* view, WifiState* state) {
+}
+
+void wifi_connect_view_clear(WifiConnectView* view) {
+    // Nothing to unsubscribe from
+}
