+++ components/tactility/src/apps/system/wifi/wifi_state.h
@@ -2,6 +2,10 @@
 
 #include <stdbool.h>
 
+#ifdef __cplusplus
+extern "C" {
+#endif
+
 typedef enum {
     WIFI_RADIO_ON,
     WIFI_RADIO_ON_PENDING,
@@ -16,3 +20,7 @@ typedef struct {
     bool scanning;
     WifiRadioState radio_state;
 } WifiState;
+
+#ifdef __cplusplus
+}
+#endif
